project create

+++ models/user.go
@@ -9,7 +9,7 @@ import (
 
 //User model definition
 type User struct {
-	ID          uint       `gorm:"primary_key" json:"-"`
+	ID          uint64       `gorm:"primary_key" json:"-"`
 	CreatedAt   time.Time  `json:"-"`
 	UpdatedAt   time.Time  `json:"-"`
 	DeletedAt   *time.Time `json:"-"`
